mito/utilities: move {SegmentedContainer} into mito/utilities
