+++ mito/tests/segmented-container/main.cc
@@ -1,0 +1,72 @@
+#include "../../utilities.h"
+
+#include <cassert>
+#include <iostream>
+
+class Simplex {
+  public:
+    Simplex(int foo) : _foo(foo), _is_valid(true) {}
+
+    void invalidate() { _is_valid = false; }
+
+    bool is_valid() { return _is_valid; }
+
+  private:
+    int _foo;
+    bool _is_valid;
+};
+
+int
+main()
+{
+    // instantiate a segmented container
+    mito::utilities::segmented_t<Simplex, 3> vector;
+
+    // assert that the container is empty and with no capacity
+    assert(vector.capacity() == 0);
+    assert(vector.size() == 0);
+
+    // add three simplices to the container
+    const auto & simplex0 = vector.add(0);
+    const auto & simplex1 = vector.add(1);
+    const auto & simplex2 = vector.add(2);
+
+    // assert that the container has 3 elements and its capacity is also 3
+    assert(vector.capacity() == 3);
+    assert(vector.size() == 3);
+
+    // erase one simplex
+    vector.erase(simplex1);
+
+    // assert that the container has 2 elements and its capacity is still 3
+    assert(vector.capacity() == 3);
+    assert(vector.size() == 2);
+
+    // add one simplex
+    const auto & simplex3 = vector.add(3);
+
+    // assert that the container has again 3 elements and its capacity is 3
+    assert(vector.capacity() == 3);
+    assert(vector.size() == 3);
+
+    // add another simplex (trigger allocation of new segment)
+    const auto & simplex4 = vector.add(4);
+
+    // assert that the container has now 4 elements and its capacity is 6
+    // (new memory allocation was in fact triggered)
+    assert(vector.capacity() == 6);
+    assert(vector.size() == 4);
+
+    // erase all the simplices
+    vector.erase(simplex0);
+    vector.erase(simplex2);
+    vector.erase(simplex3);
+    vector.erase(simplex4);
+
+    // assert that the container is empty and but has still capacity of 6
+    assert(vector.capacity() == 6);
+    assert(vector.size() == 0);
+
+    // all done
+    return 0;
+}
