+++ mito/utilities.h
@@ -1,0 +1,13 @@
+// code guard
+#if !defined(mito_utilities_h)
+#define mito_utilities_h
+
+
+// publish the interface
+// the api is in "utilities/api.h"
+#include "utilities/public.h"
+
+
+#endif
+
+// end of file
