+++ mito/utilities/externals.h
@@ -1,0 +1,15 @@
+// code guard
+#if !defined(mito_utilities_externals_h)
+#define mito_utilities_externals_h
+
+
+// externals
+#include <queue>
+#include <vector>
+#include <utility>
+#include <memory>
+
+
+#endif
+
+// end of file
