+++ mito/utilities/factories.h
@@ -1,0 +1,13 @@
+// code guard
+#if !defined(mito_utilities_factories_h)
+#define mito_utilities_factories_h
+
+
+namespace mito::utilities {
+
+}
+
+
+#endif
+
+// end of file
