+++ mito/utilities/public.h
@@ -1,0 +1,24 @@
+// code guard
+#if !defined(mito_utilities_public_h)
+#define mito_utilities_public_h
+
+
+// external packages
+#include "externals.h"
+
+// get the forward declarations
+#include "forward.h"
+
+// published type factories; this is the file you are looking for...
+#include "api.h"
+
+// classes implementation
+#include "SegmentedContainer.h"
+
+// factories implementation
+#include "factories.h"
+
+
+#endif
+
+// end of file
